Move create_clien() in the relayer
