+++ relayer/src/lib.rs
@@ -18,4 +18,5 @@ pub mod error;
 pub mod event_handler;
 pub mod event_monitor;
 pub mod store;
+pub mod tx;
 pub mod util;
